// packages/simmerbase/src/schemas/groups.ts
// Auto-generated by create-types-files.ts — 2026-02-13T00:41:02.350Z
// Do not edit manually.

import z from 'zod';

export const GroupsBaseRowSchema = z.object({
	address: z.string(),
	created_at: z.coerce.date<Date>(),
	created_by: z.uuid().nullable(),
	fax: z.string().nullable(),
	group_name: z.string(),
	id: z.uuid(),
	phone: z.string(),
	short_name: z.string().nullable(),
	updated_at: z.coerce.date<Date>(),
	updated_by: z.uuid().nullable(),
	website_url: z.string().nullable(),
});

export type GroupsBaseRow = z.infer<typeof GroupsBaseRowSchema>;

export const GroupsInsertSchema = z.object({
	address: z.string(),
	fax: z.string().nullable(),
	group_name: z.string(),
	id: z.uuid(),
	phone: z.string(),
	short_name: z.string().nullable(),
	website_url: z.string().nullable(),
});

export type GroupsInsert = z.infer<typeof GroupsInsertSchema>;

export const GroupsUpdateSchema = GroupsInsertSchema.omit({ id: true }).partial();

export type GroupsUpdate = z.infer<typeof GroupsUpdateSchema>;