// packages/simmerbase/src/schemas/units.ts
// Auto-generated by create-types-files.ts — 2026-02-13T00:41:02.350Z
// Do not edit manually.

import z from 'zod';

export const UnitsBaseRowSchema = z.object({
	abbreviation: z.string(),
	base_unit_id: z.uuid().nullable(),
	conversion_factor: z.number().nullable(),
	conversion_offset: z.number().nullable(),
	id: z.uuid(),
	unit_name: z.string(),
	unit_system: z.enum(['si', 'imperial', 'us_customary']).nullable(),
	unit_type: z.enum(['weight', 'distance', 'area', 'volume', 'temperature', 'duration', 'count', 'speed']).nullable(),
});

export type UnitsBaseRow = z.infer<typeof UnitsBaseRowSchema>;

export const UnitsInsertSchema = UnitsBaseRowSchema;

export type UnitsInsert = z.infer<typeof UnitsInsertSchema>;

export const UnitsUpdateSchema = UnitsInsertSchema.omit({ id: true }).partial();

export type UnitsUpdate = z.infer<typeof UnitsUpdateSchema>;